Ensure no warnings are generated
Update the code to ensure that the entire code base passes being built
with the options -Wall -Wextra -Werror and a few others with can be
found in CMake/CompilerWarnings.cmake.

+++ include/nuchic/Units.hh
@@ -4,46 +4,46 @@
 namespace nuchic {
     // Time literals
     constexpr double operator"" _s(long double x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _s(unsigned long long int x) {
-        return x;
+        return static_cast<double>(x);
     }
 
     // Distance literals
     constexpr double operator"" _fm(long double x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _fm(unsigned long long int x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _m(long double x) {
-        return x*1.0e15;
+        return static_cast<double>(x*1.0e15);
     }
     constexpr double operator"" _m(unsigned long long int x) {
-        return x*1.0e15;
+        return static_cast<double>(x)*1.0e15;
     }
 
     // Cross-section literals
     constexpr double operator"" _mb(long double x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _mb(unsigned long long int x) {
-        return x;
+        return static_cast<double>(x);
     }
 
     // Energy literals
     constexpr double operator"" _MeV(long double x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _MeV(unsigned long long int x) {
-        return x;
+        return static_cast<double>(x);
     }
     constexpr double operator"" _GeV(long double x) {
-        return x*1000;
+        return static_cast<double>(x*1000);
     }
     constexpr double operator"" _GeV(unsigned long long int x) {
-        return x*1000;
+        return static_cast<double>(x)*1000.0;
     }
 
 }
